discord_bot/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
yt-dlp==2023.10.13
async-timeout==4.0.3
PyNaCl==1.5.0